Debug, BSD systems refactoring ##r_debug
Separate pure BSD calls into its separate compilation unit.
